+++ Runtime/Components/DimensionOverrider.cs
@@ -7,8 +7,20 @@ namespace DimensionConverter
     public class DimensionOverrider : MonoBehaviour
     {
         [Header("Overridden on Awake")]
-        public bool is2DNot3D;
-        public bool ignoreConversionTime;
+        [SerializeField] private bool _is2DNot3D;
+        [SerializeField] private bool _ignoreConversionTime;
+
+        public bool is2DNot3D
+        {
+            get => _is2DNot3D;
+            set => _is2DNot3D = value;
+        }
+
+        public bool ignoreConversionTime
+        {
+            get => _ignoreConversionTime;
+            set => _ignoreConversionTime = value;
+        }
 
         private void Awake()
         {