+++ Runtime/Converters/CameraConverter.cs
@@ -14,40 +14,96 @@ namespace DimensionConverter
         public new Camera camera => _camera ? _camera : (_camera = GetComponent<Camera>());
         #endregion
 
+        #region Properties
+        private const float minSensorSize = 0.1f;
+
         [Tooltip("Orthographic to perspective view blends from left to right, and perspective to orthographic view blends from right to left, within range 0 to 1.")] 
-        public AnimationCurve blendCurve = AnimationCurve.EaseInOut(0f, 0f, 1f, 1f);
+        [SerializeField] private AnimationCurve _blendCurve = AnimationCurve.EaseInOut(0f, 0f, 1f, 1f);
         
         [Header("Perspective Projection")]
-        [Tooltip("The camera's view angle measured in degrees along the selected axis.")] [Range(Vector2.kEpsilon, 179f)] 
-        public float fieldOfView = 60f;
+        [Tooltip("The camera's view angle measured in degrees along the selected axis.")] [Range(Vector2.kEpsilon, 179f)]
+        [SerializeField] private float _fieldOfView = 60f;
 
-        [Tooltip("Enables Physical camera mode. When checked, the field of view is calculated from properties for simulating physical attributes (focal length, senzor size and lens shift).")] 
-        public bool usePhysicalProperties = false;
+        [Tooltip("Enables Physical camera mode. When checked, the field of view is calculated from properties for simulating physical attributes (focal length, senzor size and lens shift).")]
+        [SerializeField] private bool _usePhysicalProperties = false;
 
         [Tooltip("The simulated distance between the lens and the sensor of the physical camera. Larger values give a narrower field of view.")]
-        public float focalLength = 50f;
+        [SerializeField] private float _focalLength = 50f;
 
         [Tooltip("The size of the camera sensor in millimeters.")]
-        [Min(0.1f)] 
-        public Vector2 sensorSize = new Vector2(36f, 24f);
+        [Min(minSensorSize)]
+        [SerializeField] private Vector2 _sensorSize = new Vector2(36f, 24f);
 
         [Tooltip("Offset from the camera sensor. Use these properties to simulate a shift lens. Measured as a multiple of the sensor size.")]
-        public Vector2 lensShift = Vector2.zero;
+        [SerializeField] private Vector2 _lensShift = Vector2.zero;
 
         [Tooltip("Determines how the rendered area (resolution gate) fits within the sensor area (film gate).")]
-        public Camera.GateFitMode gateFit = Camera.GateFitMode.Horizontal;
+        [SerializeField] private Camera.GateFitMode _gateFit = Camera.GateFitMode.Horizontal;
 
         [Header("Orthographic Projection")]
-        [Tooltip("The vertical size of the camera view.")] 
-        public float orthographicSize = 5f;
+        [Tooltip("The vertical size of the camera view.")]
+        [SerializeField] private float _orthographicSize = 5f;
+
+        public AnimationCurve blendCurve
+        {
+            get => _blendCurve;
+            set => _blendCurve = value;
+        }
+
+        public float fieldOfView
+        {
+            get => _fieldOfView;
+            set => _fieldOfView = value;
+        }
+
+        public bool usePhysicalProperties
+        {
+            get => _usePhysicalProperties;
+            set => _usePhysicalProperties = value;
+        }
+
+        public float focalLength
+        {
+            get => _focalLength;
+            set => _focalLength = value;
+        }
+
+        public Vector2 sensorSize
+        {
+            get => _sensorSize;
+            set
+            {
+                _sensorSize.x = Mathf.Max(minSensorSize, value.x);
+                _sensorSize.y = Mathf.Max(minSensorSize, value.y);
+            }
+        }
+
+        public Vector2 lensShift
+        {
+            get => _lensShift;
+            set => _lensShift = value;
+        }
+
+        public Camera.GateFitMode gateFit
+        {
+            get => _gateFit;
+            set => _gateFit = value;
+        }
+
+        public float orthographicSize
+        {
+            get => _orthographicSize;
+            set => _orthographicSize = value;
+        }
 
         private Matrix4x4 perspectiveMatrix => Matrix4x4.Perspective(fieldOfView, camera.aspect, camera.nearClipPlane, camera.farClipPlane);
         private Matrix4x4 orthographicMatrix => Matrix4x4.Ortho(-orthographicSize * camera.aspect, orthographicSize * camera.aspect, -orthographicSize, orthographicSize, camera.nearClipPlane, camera.farClipPlane);
+        #endregion
 
         private void OnEnable()
         {
-            SetView(Dimension.is2DNot3D);
             Dimension.onBeforeConvert += ChangeView;
+            SetView(Dimension.is2DNot3D);
         }
 
         private void OnDisable()
@@ -103,6 +159,7 @@ namespace DimensionConverter
             if(conversionTime > 0f)
             {
                 var lerpMultiplier = 1 / conversionTime;
+                camera.orthographic = false;
 
                 while(conversionTime > 0f)
                 {