+++ Runtime/Events/RigidbodyEvent.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace DimensionConverter
+{
+    [Serializable]
+    public class RigidbodyEvent : UnityEvent<Rigidbody> { }
+}